requirements.txt
alembic==1.14.0
amqp==5.3.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
bs4==0.0.2
celery==5.4.0
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
-e git+https://Esekyi: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@github.com/Esekyi/mailSage.git@e19723feacfb062ba47a091a6b2a1b305af0374f#egg=mailsage
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.23.1
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
alembic==1.14.0
amqp==5.3.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
bs4==0.0.2
celery==5.4.0
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
-e git+https://Esekyi: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@github.com/Esekyi/mailSage.git@e19723feacfb062ba47a091a6b2a1b305af0374f#egg=mailsage
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.23.1
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
